+++ EveUtils.Client/Views/Runs/InventoryListEditorView.axaml.cs
@@ -1,0 +1,9 @@
+using Avalonia.Controls;
+using Avalonia.Markup.Xaml;
+
+namespace EveUtils.Client.Views.Runs;
+
+public partial class InventoryListEditorView : UserControl
+{
+    public InventoryListEditorView() => AvaloniaXamlLoader.Load(this);
+}